refactor: 수강생 관리 엑셀 부분 리팩터링
